Split tests into smaller files
