+++ tests/test_device/__init__.py
@@ -1,0 +1,1 @@
+"""Define tests related to the device itself."""
